refactor: split OpenAPI v3 converter
Organize conversion, normalization, validation, operation parity, JSON helpers, and atomic output into focused files without changing generated output.
